// src/types/models/dashboards/Metrics.ts
export interface SubscriptionData {
  dataOne: { number: number; value: number }[];
  dataTwo: { number: number; value: number }[];
  dataThree: { number: number; value: number }[];
}